app/requirements.txt
Django==4.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
celery==5.2.1
redis==4.0.2
google-api-python-client==2.33.0
djangorestframework==3.13.1
django-filter==21.1
Markdown==3.3.6